Refactor Register component, add random identifier generation, and update form validation
